Fixed makeBoosterArray. Also added gitignore
makeBoosterArray was defined outside of the FalconHeavy class, so when you tried to use the keyword "this" to push to the group, it did not know what this was. So I moved the functino inside the class (which means you need to get rid of the keyword "function") Then to call the function you do "this.makeBoosterArray". Also added gitignore to ignore files .idea files b/c I am using webstorm